// packages/agent-core/src/agents/runs/generated/turn-status/AgentCore/Extract/TurnStatus.ts
export type Option<A> = {
    readonly kind: "none";
} | {
    readonly kind: "some";
    readonly value: A;
};